% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mye_and_lym_methy_450k.R
\docType{data}
\name{mye_and_lym_methy_450k}
\alias{mye_and_lym_methy_450k}
\title{Leukocyte subtypes data for 450k platform}
\format{
A dataframe with 20755 genes and 48 columns
}
\usage{
data(mye_and_lym_methy_450k)
}
\description{
This beta value dataframe is a dataframe which integrated leukocyte subtypes data in GSE35069.
}
\details{
\itemize{
  \item A beta value dataframe of 20755 genes and 48 samples.
}
}
\keyword{datasets}
