% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mye_and_lym_methy_850k.R
\docType{data}
\name{mye_and_lym_methy_850k}
\alias{mye_and_lym_methy_850k}
\title{Leukocyte subtypes data for 850k platform}
\format{
A dataframe with 24350 genes and 141 columns
}
\usage{
data(mye_and_lym_methy_850k)
}
\description{
This beta value dataframe is a dataframe which integrated leukocyte subtypes data in GSE166844.
}
\details{
\itemize{
  \item A beta value dataframe of 24350 genes and 141 samples
}
}
\keyword{datasets}
