% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata_450k.R
\docType{data}
\name{refdata.450k}
\alias{refdata.450k}
\title{Reference data for 450k platform}
\format{
A dataframe with 20755 genes and 97 columns
}
\usage{
data(refdata.450k)
}
\description{
This beta value dataframe is a dataframe which integrated leukocyte subtypes data in GSE35069 and GSE88824.
}
\details{
\itemize{
  \item A beta value dataframe of 20755 genes and 96 samples
}
}
\keyword{datasets}
