% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata_850k.R
\docType{data}
\name{refdata.850k}
\alias{refdata.850k}
\title{Reference data for 850k platform}
\format{
A dataframe with 24350 genes and 142 columns
}
\usage{
data(refdata.850k)
}
\description{
This beta value dataframe is a dataframe which integrated leukocyte subtypes data in GSE166844.
}
\details{
\itemize{
  \item A beta value dataframe of 24350 genes and 141 samples
}
}
\keyword{datasets}
